refactor: RESTful API 형식으로 리팩터링
